Refactor of the entire module.
