+++ lib/Request.js
@@ -3,6 +3,8 @@
 const https = require('https');
 const url = require('url');
 
+const ENV = process.env.NODE_ENV;
+
 const REQUEST_CONFIG = {
     hostname : 'www.google-analytics.com',
     port : 443,
@@ -12,27 +14,32 @@ const REQUEST_CONFIG = {
 
 class Request {
 
-    constructor(general, session, event, pageview) {
-        this.general = url.format({ query : general});
-        this.session = url.format({ query : session});
-        this.event = url.format({ query : event});
-        this.pageview = url.format({ query : pageview});
+    constructor(req, general, session, event, pageview) {
+        this.headers = {};
+        console.log(req.headers)
+        this.headers['User-Agent'] = req.headers['user-agent'];
+
+        this.prefix = url.format({ query : general}).substring(1);
+        this.prefix += '&' + url.format({ query : session}).substring(1);
+
+        this.event = url.format({ query : event}).substring(1);
+        this.pageview = url.format({ query : pageview}).substring(1);
 
         this.setPayload();
     }
 
     setPayload() {
-        let payload = "";
+        let payload = '';
 
-        payload += this.general + '\n';
-        payload += this.session + '\n';
-        payload += this.event + '\n';
-        payload += this.pageview;
+        payload += this.prefix + '&' + this.event + '\n';
+        payload += this.prefix + '&' + this.pageview + '\n';
 
         this.payload = payload;
     }
 
     send() {
+        REQUEST_CONFIG.headers = this.headers;
+
         let request = https.request(
             REQUEST_CONFIG,
             function(response) {
@@ -42,11 +49,9 @@ class Request {
                     console.log('response=', data);
                 })
         });
-
-        request.on('error', function(e){ console.log('Error in request', e) });
+        
         request.write(this.payload);
         request.end();
-        console.log('nailed it')
     }
 }
