--- lib/environment.js
@@ -1,5 +1,0 @@
-'use strict';
-
-const ENV = process.env.NODE_ENV;
-
-
