+++ lib/models/event.js
@@ -5,7 +5,7 @@ const HIT_TYPE = 'event';
 
 class Event {
 
-    constructor(req) {
+    constructor(req, res) {
         if (!_.isObject(req)) {
             return;
         }
@@ -13,7 +13,7 @@ class Event {
         this.t = HIT_TYPE;
         this.setAction(req.method);
         this.setCategory(req);
-        this.setLabel(req.statusCode);
+        this.setLabel(res.statusCode);
     }
 
     setAction(action) {
@@ -24,9 +24,9 @@ class Event {
     }
 
     setCategory(req) {
-        this.ea = req.path;
-        if (req.route && req.route.path) {
-            this.ea = req.route.path;
+        this.ec = req.path;
+        if (req.route && req.route.path && _.isString(req.route.path)) {
+            this.ec = req.route.path;
         }
     }
