+++ lib/models/pageview.js
@@ -5,11 +5,10 @@ const HIT_TYPE = 'pageview';
 
 class Pageview {
 
-    constructor(req, options) {
+    constructor(req) {
         if (!_.isObject(req)) {
             return;
         }
-        options = options || {};
 
         this.t = HIT_TYPE;
         this.setPath(req.path);
