+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "api-ganalytics",
-  "version": "1.2.2",
+  "version": "2.0.0",
   "description": "Google Analytics for a NodeJS API.",
-  "main": "start.js",
+  "main": "index.js",
   "keywords": [
     "google",
     "analytics",
@@ -17,7 +17,7 @@
     "protocol"
   ],
   "scripts": {
-    "test": "echo \"TODO: Write tests\""
+    "test": "./node_modules/.bin/mocha --recursive"
   },
   "author": {
     "name": "Lars de Bruijn",
@@ -28,6 +28,7 @@
     "lodash": "^4.13.1"
   },
   "devDependencies": {
+    "mocha": "^2.5.3",
     "should": "^9.0.0"
   },
   "homepage": "https://github.com/ldebruijn/API-GAnalytics.git",
@@ -36,6 +37,6 @@
     "url": "https://github.com/ldebruijn/API-GAnalytics.git"
   },
   "bugs": {
-    "url": "https://github.com/ldebruijn/API-GAnalytics.git"
+    "url": "https://github.com/ldebruijn/API-GAnalytics/issues"
   }
 }
